requirements.txt
mysql-connector-python==9.1.0
requests==2.31.0
beautifulsoup4==4.12.2
boto3==1.29.0